MIPP moved to a submodule
